[UPDATE]: created merchant page and bidding approval page
