+++ src/layouts/(app)/dashboard/_components/app-sidebar.tsx
@@ -1,8 +1,7 @@
-import { memo, useState } from "react";
+import { memo } from "react";
 import {
   Sidebar,
   SidebarContent,
-  SidebarFooter,
   SidebarGroup,
   SidebarHeader,
   SidebarMenu,
@@ -10,20 +9,11 @@ import {
   SidebarMenuItem,
   SidebarRail,
 } from "@/components/ui/sidebar";
-import {
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuContent,
-  DropdownMenuItem,
-} from "@/components/ui/dropdown-menu";
 import { Link, useLocation } from "react-router-dom";
 
 import { cn } from "@/lib/utils";
-import { ChevronDown, Eye, EyeOff } from "lucide-react";
-import { useDispatch, useSelector } from "react-redux";
-import { AppDispatch, RootState } from "@/store";
-import { MdOutlineGeneratingTokens } from "react-icons/md";
-import { setSelectedToken } from "@/store/slice/wallet.slice";
+import {  useSelector } from "react-redux";
+import { RootState } from "@/store";
 
 const sidebarLinks = [
   {
@@ -52,7 +42,6 @@ const sidebarLinks = [
     ),
   },
   {
-    // label: "Property Management",
     icon: (className: string) => (
       <svg
         className={cn("!size-[25px]", className)}
@@ -73,170 +62,9 @@ const sidebarLinks = [
         />
       </svg>
     ),
-
     label: "Listings",
     path: "/listings",
-    // sub_labels: [
-    //   {
-    //     label: "Buy/Rent",
-    //     path: "/buy-rent",
-    //     icon: (className: string) => (
-    //       <svg
-    //         className={cn("!size-[18px]", className)}
-    //         viewBox="0 0 19 18"
-    //         fill="none"
-    //         xmlns="http://www.w3.org/2000/svg"
-    //       >
-    //         <path
-    //           d="M2 8.55C2 7.68127 2.1813 7.5 3.05 7.5H15.95C16.8187 7.5 17 7.68127 17 8.55V9.45C17 10.3187 16.8187 10.5 15.95 10.5H3.05C2.1813 10.5 2 10.3187 2 9.45V8.55Z"
-    //           strokeWidth="1.2"
-    //           strokeLinecap="round"
-    //         />
-    //         <path
-    //           d="M2 2.55C2 1.6813 2.1813 1.5 3.05 1.5H15.95C16.8187 1.5 17 1.6813 17 2.55V3.45C17 4.3187 16.8187 4.5 15.95 4.5H3.05C2.1813 4.5 2 4.3187 2 3.45V2.55Z"
-    //           strokeWidth="1.2"
-    //           strokeLinecap="round"
-    //         />
-    //         <path
-    //           d="M2 14.55C2 13.6813 2.1813 13.5 3.05 13.5H15.95C16.8187 13.5 17 13.6813 17 14.55V15.45C17 16.3187 16.8187 16.5 15.95 16.5H3.05C2.1813 16.5 2 16.3187 2 15.45V14.55Z"
-    //           strokeWidth="1.2"
-    //           strokeLinecap="round"
-    //         />
-    //       </svg>
-    //     ),
-    //   },
-    //   {
-    //     label: "List a Property",
-    //     path: "/new-listing",
-    //     icon: (className: string) => (
-    //       <svg
-    //         className={cn("!size-[18px]", className)}
-    //         viewBox="0 0 18 18"
-    //         fill="none"
-    //         xmlns="http://www.w3.org/2000/svg"
-    //       >
-    //         <path
-    //           d="M1.5 9C1.5 5.81802 1.5 4.22703 2.37868 3.23851C3.25736 2.25 4.67157 2.25 7.5 2.25H10.5C13.3284 2.25 14.7427 2.25 15.6213 3.23851C16.5 4.22703 16.5 5.81802 16.5 9C16.5 12.1819 16.5 13.773 15.6213 14.7615C14.7427 15.75 13.3284 15.75 10.5 15.75H7.5C4.67157 15.75 3.25736 15.75 2.37868 14.7615C1.5 13.773 1.5 12.1819 1.5 9Z"
-    //           strokeWidth="1.2"
-    //         />
-    //         <path
-    //           d="M1.5 6.75H7.5C9.6213 6.75 10.6819 6.75 11.341 7.40901C12 8.06805 12 9.1287 12 11.25V15.75"
-    //           strokeWidth="1.2"
-    //         />
-    //         <path d="M7.5 15.75V6.75" strokeWidth="1.2" />
-    //       </svg>
-    //     ),
-    //   },
-    // ],
   },
-  // {
-  //   label: "DAO  Governance",
-  //   icon: (className: string) => (
-  //     <svg
-  //       className={cn("!size-[26px]", className)}
-  //       viewBox="0 0 26 26"
-  //       fill="none"
-  //       xmlns="http://www.w3.org/2000/svg"
-  //     >
-  //       <path
-  //         d="M7.58464 2.16699C4.5931 2.16699 2.16797 4.59212 2.16797 7.58366C2.16797 9.58859 3.25726 11.3391 4.8763 12.2757V19.3304C4.8763 20.216 4.8763 20.6588 5.04123 21.057C5.20616 21.4552 5.51927 21.7683 6.1455 22.3946L7.58464 23.8337L9.86851 21.5498C9.97384 21.4445 10.0265 21.3917 10.0702 21.3346C10.185 21.1846 10.2585 21.0071 10.2835 20.8198C10.293 20.7485 10.293 20.674 10.293 20.5251C10.293 20.4045 10.293 20.3442 10.2866 20.2856C10.2698 20.132 10.2204 19.9838 10.1417 19.8509C10.1117 19.8002 10.0755 19.752 10.0031 19.6556L8.66797 17.8753L9.4263 16.8643C9.85583 16.2915 10.0706 16.0052 10.1818 15.6716C10.293 15.3381 10.293 14.9801 10.293 14.2643V12.2757C11.912 11.3391 13.0013 9.58859 13.0013 7.58366C13.0013 4.59212 10.5762 2.16699 7.58464 2.16699Z"
-  //         strokeWidth="1.625"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M7.58203 7.58301H7.59251"
-  //         strokeWidth="2.16667"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M14.082 15.167H20.582C21.5916 15.167 22.0963 15.167 22.4945 15.3319C23.0254 15.5518 23.4472 15.9736 23.6671 16.5045C23.832 16.9027 23.832 17.4074 23.832 18.417C23.832 19.4266 23.832 19.9313 23.6671 20.3295C23.4472 20.8603 23.0254 21.2822 22.4945 21.5021C22.0963 21.667 21.5916 21.667 20.582 21.667H14.082"
-  //         strokeWidth="1.625"
-  //         strokeLinecap="round"
-  //       />
-  //       <path
-  //         d="M16.25 5.41699H20.5833C21.5929 5.41699 22.0976 5.41699 22.4959 5.58192C23.0267 5.80182 23.4485 6.22362 23.6684 6.75451C23.8333 7.15269 23.8333 7.65746 23.8333 8.66699C23.8333 9.67653 23.8333 10.1813 23.6684 10.5795C23.4485 11.1103 23.0267 11.5322 22.4959 11.7521C22.0976 11.917 21.5929 11.917 20.5833 11.917H16.25"
-  //         strokeWidth="1.625"
-  //         strokeLinecap="round"
-  //       />
-  //     </svg>
-  //   ),
-  //   sub_labels: [
-  //     {
-  //       icon: (className: string) => (
-  //         <svg
-  //           className={cn("!size-[18px]", className)}
-  //           viewBox="0 0 19 18"
-  //           fill="none"
-  //           xmlns="http://www.w3.org/2000/svg"
-  //         >
-  //           <path
-  //             d="M2 8.55C2 7.68127 2.1813 7.5 3.05 7.5H15.95C16.8187 7.5 17 7.68127 17 8.55V9.45C17 10.3187 16.8187 10.5 15.95 10.5H3.05C2.1813 10.5 2 10.3187 2 9.45V8.55Z"
-  //             strokeWidth="1.2"
-  //             strokeLinecap="round"
-  //           />
-  //           <path
-  //             d="M2 2.55C2 1.6813 2.1813 1.5 3.05 1.5H15.95C16.8187 1.5 17 1.6813 17 2.55V3.45C17 4.3187 16.8187 4.5 15.95 4.5H3.05C2.1813 4.5 2 4.3187 2 3.45V2.55Z"
-  //             strokeWidth="1.2"
-  //             strokeLinecap="round"
-  //           />
-  //           <path
-  //             d="M2 14.55C2 13.6813 2.1813 13.5 3.05 13.5H15.95C16.8187 13.5 17 13.6813 17 14.55V15.45C17 16.3187 16.8187 16.5 15.95 16.5H3.05C2.1813 16.5 2 16.3187 2 15.45V14.55Z"
-  //             strokeWidth="1.2"
-  //             strokeLinecap="round"
-  //           />
-  //         </svg>
-  //       ),
-  //       label: "DOA Listings",
-  //       path: "/governance",
-  //     },
-  //     {
-  //       icon: (className: string) => (
-  //         <svg
-  //           className={cn("!size-[18px]", className)}
-  //           viewBox="0 0 18 18"
-  //           fill="none"
-  //           xmlns="http://www.w3.org/2000/svg"
-  //         >
-  //           <path
-  //             d="M1.5 9C1.5 5.81802 1.5 4.22703 2.37868 3.23851C3.25736 2.25 4.67157 2.25 7.5 2.25H10.5C13.3284 2.25 14.7427 2.25 15.6213 3.23851C16.5 4.22703 16.5 5.81802 16.5 9C16.5 12.1819 16.5 13.773 15.6213 14.7615C14.7427 15.75 13.3284 15.75 10.5 15.75H7.5C4.67157 15.75 3.25736 15.75 2.37868 14.7615C1.5 13.773 1.5 12.1819 1.5 9Z"
-  //             strokeWidth="1.2"
-  //           />
-  //           <path
-  //             d="M1.5 6.75H7.5C9.6213 6.75 10.6819 6.75 11.341 7.40901C12 8.06805 12 9.1287 12 11.25V15.75"
-  //             strokeWidth="1.2"
-  //           />
-  //           <path d="M7.5 15.75V6.75" strokeWidth="1.2" />
-  //         </svg>
-  //       ),
-  //       label: "Proposals",
-  //       path: "/proposals",
-  //     },
-  //   ],
-  //   dao_members: true,
-  // },
-  // {
-  //   icon: (className: string) => (
-  //     <svg
-  //       className={cn("!size-6", className)}
-  //       viewBox="0 0 19 19"
-  //       fill="none"
-  //       xmlns="http://www.w3.org/2000/svg"
-  //     >
-  //       <path
-  //         d="M9.50065 17.4168C13.8729 17.4168 17.4173 13.8724 17.4173 9.50016C17.4173 5.12791 13.8729 1.5835 9.50065 1.5835C5.1284 1.5835 1.58398 5.12791 1.58398 9.50016C1.58398 13.8724 5.1284 17.4168 9.50065 17.4168Z"
-  //         strokeWidth="1.2"
-  //       />
-  //       <path
-  //         d="M5.9375 13.4582C7.78343 11.5248 11.1967 11.4337 13.0625 13.4582M11.4753 7.52067C11.4753 8.61372 10.5879 9.49984 9.49327 9.49984C8.39871 9.49984 7.51131 8.61372 7.51131 7.52067C7.51131 6.42761 8.39871 5.5415 9.49327 5.5415C10.5879 5.5415 11.4753 6.42761 11.4753 7.52067Z"
-  //         strokeWidth="1.2"
-  //         strokeLinecap="round"
-  //       />
-  //     </svg>
-  //   ),
-  //   label: "Profile",
-  //   path: "/profile",
-  // },
   {
     icon: (className: string) => (
       <svg
@@ -303,65 +131,28 @@ const sidebarLinks = [
     ),
     label: "Trading",
     path: "/trading",
-  },
-  // {
-  //   icon: (className: string) => (
-  //     <svg
-  //       className={cn("!size-6", className)}
-  //       viewBox="0 0 24 24"
-  //       fill="none"
-  //       xmlns="http://www.w3.org/2000/svg"
-  //     >
-  //       <path
-  //         d="M10 14H3V21H10V14Z"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M10 3H3V10H10V3Z"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M14 4H21"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M14 9H21"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M14 15H21"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //       <path
-  //         d="M14 20H21"
-  //         strokeWidth="2"
-  //         strokeLinecap="round"
-  //         strokeLinejoin="round"
-  //       />
-  //     </svg>
-  //   ),
-  //   label: "Transaction history",
-  //   path: "/transaction-history",
-  // },
+    },
+    {
+        label: "New Listings",
+        path: "/new-listings",
+        icon: (className: string) => (
+            <svg className={cn("!size-[26px]", className)} viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M10 14.5H3V21.5H10V14.5Z" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M10 3.5H3V10.5H10V3.5Z" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M14 4.5H21" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M14 9.5H21" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M14 15.5H21" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M14 20.5H21" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+            </svg>
+        ),
+    },
 ];
 
 const AppSidebar = ({ ...props }: React.ComponentProps<typeof Sidebar>) => {
-  const dispatch = useDispatch<AppDispatch>();
-  const walletState = useSelector((state: RootState) => state.wallet);
+//   const dispatch = useDispatch<AppDispatch>();
+//   const walletState = useSelector((state: RootState) => state.wallet);
 
-
-
-  const [shouldShowBalance, setShouldShowBalance] = useState(true);
+//   const [shouldShowBalance, setShouldShowBalance] = useState(true);
 
   return (
     <Sidebar {...props}>
@@ -378,7 +169,7 @@ const AppSidebar = ({ ...props }: React.ComponentProps<typeof Sidebar>) => {
       <SidebarContent>
         <NavMain />
       </SidebarContent>
-      {walletState?.isWalletConnected && (
+      {/* {walletState?.isWalletConnected && (
         <SidebarFooter>
           <div className="flex flex-col">
             <div className="flex items-center justify-between border-t border-[#EAECF0] px-8 py-4">
@@ -455,7 +246,7 @@ const AppSidebar = ({ ...props }: React.ComponentProps<typeof Sidebar>) => {
             </div>
           </div>
         </SidebarFooter>
-      )}
+      )} */}
       <SidebarRail />
     </Sidebar>
   );
@@ -483,9 +274,9 @@ export function NavMain() {
               <SidebarMenuButton
                 asChild
                 className={cn(
-                  "relative h-14 rounded-none rounded-l-full p-0 pl-6 transition hover:bg-secondary",
+                  "relative h-14 font-medium rounded-none rounded-l-full p-0 pl-6 transition hover:bg-secondary",
                   {
-                    "bg-[#e7fefc] font-medium text-primary hover:bg-[#dbfffc] hover:text-primary":
+                    "bg-[#e7fefc] text-primary hover:bg-[#dbfffc] hover:text-primary":
                       isActive(path),
                   },
                 )}
@@ -505,7 +296,7 @@ export function NavMain() {
                         ? "stroke-[#056F67]"
                         : "stroke-muted-foreground group-hover:stroke-foreground transition-[stroke]",
                     )}
-                  <span className="text-base md:text-lg">{label}</span>
+                  <span className="text-base tracking-wide">{label}</span>
                   {isActive(path) && (
                     <span className="absolute right-0 top-0 h-full w-1 bg-primary" />
                   )}
