+++ src/pages/(app)/dashboard/dashboard.page.tsx
@@ -29,7 +29,6 @@ export default function DashboardPage() {
 
   return (
     <div className="flex flex-col gap-4 py-4">
-      {/* <button onClick={chii}>click</button> */}
       <div className="h-[200px] w-full rounded-2xl bg-gradient-to-l from-[#0D857C] to-[#0EB9AC] p-[1px] sm:h-[240px] md:rounded-3xl">
         <div className="flex size-full overflow-hidden rounded-[inherit] bg-gradient-to-r from-[#056F67] to-[#0AADA1] text-white">
           <div className="lg:px-12">
