+++ src/pages/(app)/list-property/_components/indicator.tsx
