+++ src/pages/(app)/list-property/steps/building-steps/amenities.tsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import {
   buildingFormSchema,
   BuildingFormSchemaTypes,
-} from "../../new-listing.page";
+} from "../../list-property.page";
 import { zodResolver } from "@hookform/resolvers/zod";
 import {
   setCurrentStep,
