+++ src/pages/(app)/list-property/steps/land-step/survey-plan.tsx
@@ -1,88 +1,120 @@
 import { AppDispatch, RootState } from "@/store";
+import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { useDispatch, useSelector } from "react-redux";
-import {
-  buildingFormSchema,
-  BuildingFormSchemaTypes,
-} from "../../new-listing.page";
-import { zodResolver } from "@hookform/resolvers/zod";
+import { landFormSchema, LandFormSchemaTypes } from "../../list-property.page";
 import {
   setCurrentStep,
   updateFormData,
 } from "@/store/slice/new-listing.slice";
-import { FormControl, FormField, FormItem } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
 import { cn } from "@/lib/utils";
 import { Separator } from "@/components/ui/separator";
 import { Button } from "@/components/ui/button";
+import { ArrowDown, UploadIcon, Paperclip } from "lucide-react";
 import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { ArrowDown, CalendarIcon } from "lucide-react";
-import { Calendar } from "@/components/ui/calendar";
-import { format } from "date-fns";
+  FileInput,
+  FileUploader,
+  FileUploaderContent,
+  FileUploaderItem,
+} from "@/components/extension/file-uploader";
+import { DropzoneOptions } from "react-dropzone";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import { Input } from "@/components/ui/input";
+import { Textarea } from "@/components/ui/textarea";
 
-export default function DetailsForm() {
+
+export default function SurveyPlanForm() {
   const dispatch = useDispatch<AppDispatch>();
   const formData = useSelector((state: RootState) => state.newListing.formData);
 
-  const form = useForm<BuildingFormSchemaTypes>({
+  const form = useForm<LandFormSchemaTypes>({
     resolver: zodResolver(
-      buildingFormSchema.pick({
-        // rangeFrom: true,
-        // rangeTo: true,
-        price: true,
-        rooms: true,
-        bathrooms: true,
-        bedrooms: true,
-        yearBuilt: true,
-        structureType: true,
-        propertySize: true,
+      landFormSchema.pick({
+        landSize: true,
+        // boundariesFrom: true,
+        // boundariesTo: true,
+        surveyDescription: true,
+        surveyPlan: true,
+        // propertySize: true,
+        // yearBuilt: true,
+        // plotSize: true,
       }),
     ),
-    defaultValues: {
-      ...formData,
-    },
+    defaultValues: formData,
   });
 
   const {
     formState: { errors },
   } = form;
 
-  const onSubmit = (data: Partial<BuildingFormSchemaTypes>) => {
+  const onSubmit = async (data: Partial<LandFormSchemaTypes>) => {
     dispatch(updateFormData(data));
-    dispatch(setCurrentStep(4));
+    dispatch(setCurrentStep(3));
   };
 
   return (
     <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-col">
       <div className="flex flex-col gap-3">
+        <FormField
+          control={form.control}
+          name="landSize"
+          render={({ field }) => (
+            <FormItem className="w-full">
+              <FormControl>
+                <Input
+                  placeholder="Land size(sqm)"
+                  type="number"
+                  className={cn("text-foreground", {
+                    "border-red-500 focus-visible:ring-red-500":
+                      errors.landSize?.message,
+                  })}
+                  {...field}
+                />
+              </FormControl>
+              {errors.landSize && (
+                <p
+                  className={cn(
+                    "text-sm font-medium text-red-500 dark:text-red-900",
+                  )}
+                >
+                  {errors.landSize?.message}
+                </p>
+              )}
+            </FormItem>
+          )}
+        />
+
         {/* <div className="flex flex-col items-center gap-4 md:flex-row">
           <FormField
             control={form.control}
-            name="rangeFrom"
+            name="boundariesFrom"
             render={({ field }) => (
               <FormItem className="w-full">
                 <FormControl>
                   <Input
-                    placeholder="Price Range (from)"
+                    placeholder="Boundaries (from)"
                     type="number"
                     className={cn("text-foreground", {
                       "border-red-500 focus-visible:ring-red-500":
-                        errors.rangeFrom?.message,
+                        errors.boundariesFrom?.message,
                     })}
                     {...field}
                   />
                 </FormControl>
-                {errors.rangeFrom && (
+                {errors.boundariesFrom && (
                   <p
                     className={cn(
                       "text-sm font-medium text-red-500 dark:text-red-900",
                     )}
                   >
-                    {errors.rangeFrom?.message}
+                    {errors.boundariesFrom?.message}
                   </p>
                 )}
               </FormItem>
@@ -90,140 +122,152 @@ export default function DetailsForm() {
           />
           <FormField
             control={form.control}
-            name="rangeTo"
+            name="boundariesTo"
             render={({ field }) => (
               <FormItem className="w-full">
                 <FormControl>
                   <Input
-                    placeholder="Price Range (to)"
+                    placeholder="Boundaries (to)"
                     type="number"
                     className={cn("text-foreground", {
                       "border-red-500 focus-visible:ring-red-500":
-                        errors.rangeTo?.message,
+                        errors.boundariesTo?.message,
                     })}
                     {...field}
                   />
                 </FormControl>
-                {errors.rangeTo && (
+                {errors.boundariesTo && (
                   <p
                     className={cn(
                       "text-sm font-medium text-red-500 dark:text-red-900",
                     )}
                   >
-                    {errors.rangeTo?.message}
+                    {errors.boundariesTo?.message}
                   </p>
                 )}
               </FormItem>
             )}
           />
         </div> */}
+
         <FormField
           control={form.control}
-          name="price"
+          name="surveyDescription"
           render={({ field }) => (
             <FormItem className="w-full">
               <FormControl>
-                <Input
-                  placeholder="Price"
-                  type="number"
-                  className={cn("text-foreground", {
-                    "border-red-500 focus-visible:ring-red-500":
-                      errors.price?.message,
-                  })}
+                <Textarea
+                  placeholder="Floor Plan Descriptions"
                   {...field}
+                  className={cn("h-[144px] resize-none py-4 text-foreground", {
+                    "border-red-500 focus-visible:ring-red-500":
+                      errors.surveyDescription?.message,
+                  })}
                 />
               </FormControl>
-              {errors.rooms && (
+              {errors.surveyDescription && (
                 <p
                   className={cn(
                     "text-sm font-medium text-red-500 dark:text-red-900",
                   )}
                 >
-                  {errors.price?.message}
+                  {errors.surveyDescription.message}
                 </p>
               )}
             </FormItem>
           )}
         />
+
         <FormField
           control={form.control}
-          name="rooms"
+          name="surveyPlan"
           render={({ field }) => (
-            <FormItem className="w-full">
-              <FormControl>
-                <Input
-                  placeholder="Room(s)"
-                  type="number"
-                  className={cn("text-foreground", {
-                    "border-red-500 focus-visible:ring-red-500":
-                      errors.rooms?.message,
-                  })}
-                  {...field}
-                />
-              </FormControl>
-              {errors.rooms && (
-                <p
+            <FormItem>
+              <FileUploader
+                value={field.value}
+                onValueChange={field.onChange}
+                dropzoneOptions={
+                  {
+                    accept: {
+                      "image/*": [".jpg", ".jpeg", ".png", ".webp"],
+                    },
+                    maxFiles: 10,
+                    maxSize: 1024 * 1024 * 5,
+                    multiple: true,
+                  } satisfies DropzoneOptions
+                }
+                className="relative"
+              >
+                <FileInput
                   className={cn(
-                    "text-sm font-medium text-red-500 dark:text-red-900",
+                    "overflow-hidden rounded-md border border-neutral-200 sm:rounded-xl",
+                    {
+                      "border-red-500": errors.surveyPlan,
+                    },
                   )}
                 >
-                  {errors.rooms?.message}
-                </p>
-              )}
+                  <div className="flex flex-col gap-4 bg-background p-6">
+                    <FormLabel className="text-base font-normal text-muted-foreground">
+                      Image of survey Plan
+                    </FormLabel>
+
+                    <div className="flex aspect-video w-full flex-col items-center justify-center gap-4 rounded-xl border border-dashed border-primary/50 bg-[#F8FAFA]">
+                      <UploadIcon className="size-10 text-primary/80" />
+                      <p className="text-primary">
+                        Drag and Drop file here or{" "}
+                        <span className="underline">Choose file</span>
+                      </p>
+                    </div>
+                  </div>
+                </FileInput>
+                <FileUploaderContent>
+                  {field.value && field.value.length > 0 && (
+                    <ScrollArea className="mt-2 max-h-36 rounded-xl border bg-background px-1 py-2">
+                      {field.value.map((file, i) => (
+                        <FileUploaderItem
+                          key={i}
+                          index={i}
+                          className="!rounded-md"
+                          type="document"
+                        >
+                          <Paperclip className="size-5" />
+                          <span className="text-sm font-medium text-foreground">
+                            {file.name}
+                          </span>
+                        </FileUploaderItem>
+                      ))}
+                    </ScrollArea>
+                  )}
+                </FileUploaderContent>
+              </FileUploader>
+              <FormMessage />
             </FormItem>
           )}
         />
-        <FormField
+
+        {/* <FormField
           control={form.control}
-          name="bedrooms"
+          name="propertySize"
           render={({ field }) => (
             <FormItem className="w-full">
               <FormControl>
                 <Input
-                  placeholder="Bedroom(s)"
+                  placeholder="Property Sizes (Square meters)"
                   type="number"
                   className={cn("text-foreground", {
                     "border-red-500 focus-visible:ring-red-500":
-                      errors.bedrooms?.message,
+                      errors.propertySize?.message,
                   })}
                   {...field}
                 />
               </FormControl>
-              {errors.bedrooms && (
+              {errors.propertySize && (
                 <p
                   className={cn(
                     "text-sm font-medium text-red-500 dark:text-red-900",
                   )}
                 >
-                  {errors.bedrooms?.message}
-                </p>
-              )}
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="bathrooms"
-          render={({ field }) => (
-            <FormItem className="w-full">
-              <FormControl>
-                <Input
-                  placeholder="Bathroom(s)"
-                  type="number"
-                  className={cn("text-foreground", {
-                    "border-red-500 focus-visible:ring-red-500":
-                      errors.bathrooms?.message,
-                  })}
-                  {...field}
-                />
-              </FormControl>
-              {errors.bathrooms && (
-                <p
-                  className={cn(
-                    "text-sm font-medium text-red-500 dark:text-red-900",
-                  )}
-                >
-                  {errors.bathrooms?.message}
+                  {errors.propertySize?.message}
                 </p>
               )}
             </FormItem>
@@ -287,60 +331,32 @@ export default function DetailsForm() {
 
         <FormField
           control={form.control}
-          name="propertySize"
+          name="plotSize"
           render={({ field }) => (
             <FormItem className="w-full">
               <FormControl>
                 <Input
-                  placeholder="Property Sizes (Square meters)"
+                  placeholder="Property plot size"
                   type="number"
                   className={cn("text-foreground", {
                     "border-red-500 focus-visible:ring-red-500":
-                      errors.propertySize?.message,
+                      errors.plotSize?.message,
                   })}
                   {...field}
                 />
               </FormControl>
-              {errors.propertySize && (
+              {errors.plotSize && (
                 <p
                   className={cn(
                     "text-sm font-medium text-red-500 dark:text-red-900",
                   )}
                 >
-                  {errors.propertySize?.message}
+                  {errors.plotSize?.message}
                 </p>
               )}
             </FormItem>
           )}
-        />
-        <FormField
-          control={form.control}
-          name="structureType"
-          render={({ field }) => (
-            <FormItem className="w-full">
-              <FormControl>
-                <Input
-                  placeholder="Structure Type"
-                  type="text"
-                  className={cn("text-foreground", {
-                    "border-red-500 focus-visible:ring-red-500":
-                      errors.structureType?.message,
-                  })}
-                  {...field}
-                />
-              </FormControl>
-              {errors.structureType && (
-                <p
-                  className={cn(
-                    "text-sm font-medium text-red-500 dark:text-red-900",
-                  )}
-                >
-                  {errors.structureType?.message}
-                </p>
-              )}
-            </FormItem>
-          )}
-        />
+        /> */}
       </div>
 
       <Separator className="my-6 h-px w-full" />
@@ -351,7 +367,7 @@ export default function DetailsForm() {
           size={"lg"}
           variant={"outline"}
           className="w-[100px] rounded-full"
-          onClick={() => dispatch(setCurrentStep(2))}
+          onClick={() => dispatch(setCurrentStep(1))}
         >
           Back
         </Button>
