+++ src/pages/(app)/list-property/steps/property-type-form.tsx
