+++ src/pages/index.tsx
@@ -1,30 +1,28 @@
+import BidPage from "./(app)/bid/bid.page";
 import BuyOrRentPage from "./(app)/buy-or-rent/buy-or-rent.page";
 import DashboardPage from "./(app)/dashboard/dashboard.page";
-// import GovernancePage from "./(app)/governance/governance.page";
-import NewListingPage from "./(app)/new-listing/new-listing.page";
+import ListPropertyPage from "./(app)/list-property/list-property.page";
+import NewListingsPage from "./(app)/new-listings/new-listings.page";
 import GetVerifiedPage from "./(app)/onboarding/account-type/get-verified.page";
 import PropertyManagementPage from "./(app)/onboarding/account-type/property-management.page";
 import OnboardingPage from "./(app)/onboarding/onboarding.page";
 import ProfilePage from "./(app)/profile/profile.page";
 import PropertyDetailsPage from "./(app)/property-details/property-details.page";
-// import NewProposalPage from "./(app)/proposal/new-proposal.page";
-// import ProposalsPage from "./(app)/proposal/proposals.page";
 import TradingPage from "./(app)/trading/trading.page";
 import HomePage from "./(landing)/home.page";
 
 export const pages = {
   dashboardPage: <DashboardPage />,
   profilePage: <ProfilePage />,
-  listPropertyPage: <NewListingPage />,
+  listPropertyPage: <ListPropertyPage />,
+  newListingsPage: <NewListingsPage />,
   buyOrRentPage: <BuyOrRentPage />,
-  // governancePage: <GovernancePage />,
-  // proposalsPage: <ProposalsPage />,
-  // newProposalPage: <NewProposalPage />,
   tradingPage: <TradingPage />,
-  propertyDetailsPage: <PropertyDetailsPage />,
+    propertyDetailsPage: <PropertyDetailsPage />,
+  approveBidPage: <BidPage />,
 
   onboardingPage: <OnboardingPage />,
-  // // onboardings
+  // onboardings
   propertyManagementPage: <PropertyManagementPage />,
   getVerifiedPage: <GetVerifiedPage />,
   // landing page
